moved legacy fixes code from utils to fixes
