--- src/rook/utils/data_utils/__init__.py
@@ -1,1 +1,0 @@
-"""Dataset utility helpers used by rook fix routines."""
